Apply PR CR suggestions
