+++ tests/workflows/integration_tests/execution/test_workflow_with_masked_crop.py
@@ -1,4 +1,3 @@
-import matplotlib.pyplot as plt
 import numpy as np
 
 from inference.core.env import WORKFLOWS_MAX_CONCURRENT_STEPS
@@ -33,10 +32,11 @@ MASKED_CROP_WORKFLOW = {
             "confidence": "$inputs.confidence",
         },
         {
-            "type": "roboflow_core/masked_crop@v1",
+            "type": "roboflow_core/dynamic_crop@v1",
             "name": "cropping",
             "image": "$inputs.image",
             "predictions": "$steps.segmentation.predictions",
+            "mask_opacity": 1.0,
         },
     ],
     "outputs": [
@@ -44,6 +44,11 @@ MASKED_CROP_WORKFLOW = {
             "type": "JsonField",
             "name": "crops",
             "selector": "$steps.cropping.crops",
+        },
+        {
+            "type": "JsonField",
+            "name": "predictions",
+            "selector": "$steps.segmentation.predictions",
         },
     ],
 }
@@ -87,8 +92,21 @@ def test_workflow_with_masked_crop(
     assert len(result) == 1, "Expected 1 element in the output for one input image"
     assert set(result[0].keys()) == {
         "crops",
+        "predictions",
     }, "Expected all declared outputs to be delivered"
     assert len(result[0]["crops"]) == 2, "Expected 2 crops for two dogs detected"
+    crop_image = result[0]["crops"][0].numpy_image
+    (x_min, y_min, x_max, y_max) = (
+        result[0]["predictions"].xyxy[0].round().astype(dtype=int)
+    )
+    crop_mask = result[0]["predictions"].mask[0][y_min:y_max, x_min:x_max]
+    pixels_outside_mask = np.where(
+        np.stack([crop_mask] * 3, axis=-1) == 0,
+        crop_image,
+        np.zeros_like(crop_image),
+    )
+    pixels_sum = pixels_outside_mask.sum()
+    assert pixels_sum == 0, "Expected everything black outside mask"
 
 
 def test_workflow_with_masked_crop_when_nothing_gets_predicted(
@@ -117,5 +135,6 @@ def test_workflow_with_masked_crop_when_nothing_gets_predicted(
     assert len(result) == 1, "Expected 1 element in the output for one input image"
     assert set(result[0].keys()) == {
         "crops",
+        "predictions",
     }, "Expected all declared outputs to be delivered"
     assert len(result[0]["crops"]) == 0, "Expected 0 crops detected"
